+++ tests/unit/test_client.py
@@ -14,8 +14,8 @@ from finetuner.constants import (
     SYNTHESIS_TASK,
     TRAINING_TASK,
 )
-from finetuner.data import DATA_SYNTHESIS_EN
 from finetuner.experiment import Experiment
+from finetuner.model import synthesis_models_en
 
 
 def test_create_experiment(client_mocker, name='name'):
@@ -99,7 +99,7 @@ def test_create_synthesis_run(client_mocker, experiment_name='exp', run_name='ru
     config = Experiment._create_synthesis_config(
         query_data='query_data_name',
         corpus_data='corpus_data_name',
-        models=DATA_SYNTHESIS_EN,
+        models=synthesis_models_en,
         num_relations=3,
         experiment_name=experiment_name,
         run_name=run_name,
