+++ tests/unit/test_experiment.py
@@ -51,8 +51,8 @@ from finetuner.constants import (
     TRAIN_DATA,
     VAL_SPLIT,
 )
-from finetuner.data import DATA_SYNTHESIS_EN
 from finetuner.experiment import Experiment
+from finetuner.model import synthesis_models_en
 
 
 @pytest.fixture
@@ -117,14 +117,14 @@ def test_create_synthesis_run(experiment):
     run = experiment.create_synthesis_run(
         query_data=query_data,
         corpus_data=corpus_data,
-        models=DATA_SYNTHESIS_EN,
+        models=synthesis_models_en,
         num_relations=3,
         run_name=run_name,
     )
     expected_config = Experiment._create_synthesis_config(
         query_data=query_data_name,
         corpus_data=corpus_data_name,
-        models=DATA_SYNTHESIS_EN,
+        models=synthesis_models_en,
         num_relations=3,
         experiment_name=experiment.name,
         run_name=run_name,
@@ -220,10 +220,10 @@ def test_create_synthesis_run_config():
             CORPUS: 'corpus_data',
         },
         RELATION_MINING: {
-            MODELS: [DATA_SYNTHESIS_EN.relation_miner],
+            MODELS: [synthesis_models_en.relation_miner],
             NUM_RELATIONS: 3,
         },
-        CROSS_ENCODER: DATA_SYNTHESIS_EN.cross_encoder,
+        CROSS_ENCODER: synthesis_models_en.cross_encoder,
         MAX_NUM_DOCS: None,
         EXPERIMENT_NAME: 'exp name',
         PUBLIC: False,
@@ -233,7 +233,7 @@ def test_create_synthesis_run_config():
     config = Experiment._create_synthesis_config(
         train_data='train_data',
         experiment_name='exp name',
-        models=DATA_SYNTHESIS_EN,
+        models=synthesis_models_en,
         run_name='run name',
         query_data='query_data',
         corpus_data='corpus_data',
